move profiling onboarding to utils
