+++ static/app/gettingStartedDocs/python/sanic.tsx
@@ -12,9 +12,9 @@ import {
 import {
   crashReportOnboardingPython,
   featureFlagOnboarding,
-  getPythonProfilingOnboarding,
 } from 'sentry/gettingStartedDocs/python/python';
 import {t, tct} from 'sentry/locale';
+import {getPythonProfilingOnboarding} from 'sentry/utils/gettingStartedDocs/python';
 
 type Params = DocsParams;
